[#13] Refactor: Main Page 비지니스 로직 리팩토링
